task add pr

+++ 0x0A-argc_argv/3-mul.c
@@ -14,7 +14,7 @@ int main(int argc, char *argv[])
 
 	if (argc == 3)
 	{
-		s = (int)atoi(argv[1]) * (int)atoi(argv[2]);
+		s = atoi(argv[1]) * atoi(argv[2]);
 		printf("%d\n", s);
 	}
 	else
